Renamed commands to have underscores in them.
importrules -> import_rules
reloadrules -> reload_rules

Deprecated all commands.
